+++ src/dashboard/CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to
 - add ProConnect authentication system
 - add dashboard homepage
 - add consent form to manage consents of one or many entities
+- added a validated consent page allowing consultation of validated consent for the 
+  current period.
 - add an email notification to users (via Brevo) after they have validated their consents.
 - add admin integration for Entity, DeliveryPoint and Consent
 - add mass admin action (make revoked) for consents